chore: refactor local docker executor, deleting manage_jobs.py
Now that the old path is gone, we can move things around to where they
belong. Main goal is to move all local-docker related things into the
local docker executor module.
